Insert statements for 2 of 4 tables
I managed to write two insert statements for tables customer and
appliance. I also added two functions to the menu module, basic stuff.
the nest step is to pull the id number from the customer and relate it
to an appliance. I also need to check the database for customers who are
already in the DB. Then I can bypass the insert statement and go on to
add the data to the sales and sold_appl tables.