developed search/sale functions
For database.py, I have added some functions for searching the database,
select statements by name, and phone number. Also the sale of an item
has a lot of working parts. I have all of the tables storing data, and I
added a menu feature to the menu.py module for searching. Saleticket1
and 2 are the backbone to most of the data transfer in the program thus
far. I used the "re" module to pick out multiple appliance ids for the
purchase of more that one. i think i need some foreign keys to create
some integrity in my tables.