pgRouting extension
I have been experimenting with the pgRouting extension to store rout
data for customers, as well as a way to quickly compare directions from
the customer to a rout. The data i have needs to be prepared before it
can be used to calculate routs. This is still the beginning stage.