Finished the money table, for now...
I was able to get the string formatting to work by converting the string
to a float, line 112 and 118. I used the postgres conversion style
because everything I tried would give a syntax error. The money table
can now be updated neatly, and it creates a new row every time the
current day is different than what is in there. Everything works with
the totals too, but i will need to give consideration to totals being
input at the point of sale, not when the appliance is input into the
database. Every situation can be different when receiving funds. Also,
the date will be the key in this table.